Web_shargpt/backend_requirements.txt
# NL2SQL自动化数据生成系统 - 后端依赖

# FastAPI框架和服务器
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0

# 数据库连接
pymysql>=1.1.0
psycopg2-binary>=2.9.0
pyodbc>=4.0.0
cryptography>=41.0.0  # MySQL 8.0+ 认证支持

# LLM客户端
openai>=1.0.0

# SQL解析与验证
sqlglot>=20.0.0

# 配置文件
pyyaml>=6.0

# 工具库
requests>=2.31.0
aiofiles>=23.0.0
python-dotenv>=1.0.0

# 异步任务和进度跟踪
asyncio
